style: :lipstick: Renamed component style files
Issue: #152
