+++ src/components/MusicList/SearchForm.tsx
@@ -3,7 +3,7 @@ import { useState } from "react";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faSearch } from "@fortawesome/free-solid-svg-icons";
 
-import styles from "./style.module.css";
+import styles from "./styles.module.css";
 
 export type SearchData = {
   title: string;
